Rename depth camera to fit base
