+++ robot_calibration/CMakeLists.txt
@@ -108,7 +108,7 @@ add_dependencies(robot_calibration robot_calibration_msgs_generate_messages_cpp)
 add_library(robot_calibration_feature_finders src/finders/checkerboard_finder.cpp
                                               src/finders/led_finder.cpp
                                               src/finders/rgb_camera_manager.cpp
-                                              src/finders/depth_camera.cpp
+                                              src/finders/depth_camera_manager.cpp
                                               src/finders/plane_finder.cpp)
 target_link_libraries(robot_calibration_feature_finders robot_calibration
                                                         ${Boost_LIBRARIES}
