+++ robot_calibration/include/robot_calibration/capture/checkerboard_finder.h
@@ -21,7 +21,7 @@
 #ifndef ROBOT_CALIBRATION_CAPTURE_CHECKERBOARD_FINDER_H
 #define ROBOT_CALIBRATION_CAPTURE_CHECKERBOARD_FINDER_H
 
-#include <robot_calibration/capture/depth_camera.h>
+#include <robot_calibration/capture/depth_camera_manager.h>
 #include <robot_calibration/capture/camera_manager_base.h>
 #include <robot_calibration/plugins/feature_finder.h>
 #include <robot_calibration_msgs/CalibrationData.h>
