+++ robot_calibration/include/robot_calibration/capture/depth_camera_manager.h
