+++ robot_calibration/include/robot_calibration/capture/plane_finder.h
@@ -18,7 +18,7 @@
 #define ROBOT_CALIBRATION_CAPTURE_PLANE_FINDER_H
 
 #include <ros/ros.h>
-#include <robot_calibration/capture/depth_camera.h>
+#include <robot_calibration/capture/depth_camera_manager.h>
 #include <robot_calibration/plugins/feature_finder.h>
 #include <robot_calibration_msgs/CalibrationData.h>
 #include <cv_bridge/cv_bridge.h>
